backend/requirements.txt
# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=1.4.0

# Academic Paper Sources
arxiv>=1.4.7

# Social Media APIs
tweepy>=4.14.0

# Text Processing and Summarization
transformers>=4.35.0
torch>=2.0.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2

# Typing
typing-extensions>=4.7.0
